Add files via upload
This is a single page application which allows a user to type in a short task description and a due date and add it to a list of tasks.
The list of tasks is displayed upon loading the page and is updated dynamically when a user adds a new task. Each task has the capability to be edited, marked as complete, and deleted.